backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
neo4j==5.27.*
anthropic==0.49.*
openai==1.*
ollama==0.*
boto3>=1.35
python-multipart==0.0.20
pydantic==2.10.*
python-dotenv==1.0.*
pyyaml>=6.0
python-pptx>=0.6.23
python-docx>=1.1.0
